requirements.txt
asgiref==3.7.2
bcc==0.18.0
blinker==1.8.2
Brlapi==0.8.3
certifi==2020.6.20
channels==4.1.0
chardet==4.0.0
chrome-gnome-shell==0.0.0
click==8.1.7
colorama==0.4.4
command-not-found==0.3
contourpy==1.2.1
crispy-bootstrap5==2023.10
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dash==2.9.3
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
Django==4.2.9
django-admin-datta==1.0.10
django-api-generator==1.0.17
django-crispy-forms==2.1
django-debug-toolbar==4.2.0
django-dynamic-datatb==1.0.25
django-plotly-dash==2.3.1
djangorestframework==3.15.2
dpd-components==0.1.0
Flask==3.0.3
fonttools==4.53.0
gunicorn==21.2.0
hidpidaemon==18.4.6
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.4
keyring==23.5.0
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
macaroonbakery==1.3.1
MarkupSafe==2.1.5
matplotlib==3.9.0
more-itertools==8.10.0
netaddr==0.8.0
netifaces==0.11.0
numpy==2.0.0
oauthlib==3.2.0
packaging==24.1
pandas==2.2.2
pillow==10.3.0
plotly==5.22.0
pop-transition==1.1.2
protobuf==3.12.4
pycairo==1.20.1
pycups==2.0.1
pydbus==0.6.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-gnupg==0.4.8
python-xlib==0.29
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
repolib==2.2.1
repoman==1.4.0
requests==2.25.1
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
sessioninstaller==0.0.0
six==1.16.0
sqlparse==0.4.4
str2bool==1.1
systemd-python==234
tenacity==8.4.2
typing_extensions==4.9.0
tzdata==2023.4
ubuntu-drivers-common==0.0.0
ufw==0.36.1
urllib3==1.26.5
wadllib==1.3.6
Werkzeug==3.0.3
whitenoise==6.6.0
xdg==5
xkit==0.0.0
zipp==1.0.0
